pyramid pattern
This C program is designed to print a pattern of a specified symbol, entered by the user, in the form of a pyramid. The user is prompted to input the number of rows and the symbol they want to use for the pattern. The program then utilizes nested loops to create the desired pattern.